requirements.txt
aiodocker
django
django-allauth[socialaccount]
markdown-it-py[plugins]
